Copy environment to subprocess
If cgat is installed in a venv, the location of the "cgat" wrapper script
is added to PATH when the venv is sourced.

venvs are typically sourced in the user .bashrc.

.bashrc is not evaluated by non-interactive shells.

The shell opened by subprocess with "shell=True" is non-interactive.

Thus, unless the "cgat" command is on some path not set in .bashrc, execution of
jobs locally (i.e. non cluster) that use the cgat wrapper (e.g. cgat csv2db)
will fail for venvs.

This fixes the issue by copying the environment path to the subprocess
shell.